# apps/backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lead Cat API",
    "version": "1.0.0",
    "description": "Mini App-facing endpoints and health check for frontend codegen."
  },
  "paths": {
...
    "/api/miniapp/meetings": {
      "get": {
        "operationId": "miniapp_meetings_get",
        "tags": [
          "miniapp"
        ],
        "summary": "List authed user meetings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "upcoming",
                "past",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meetings list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniAppMeetingsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "miniappCreateMeeting",
        "tags": [
          "miniapp"
        ],
        "summary": "Create a meeting (once or recurring series)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppMeetingCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created meeting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meeting"
                  ],
                  "properties": {
                    "meeting": {
                      "$ref": "#/components/schemas/MiniAppMeeting"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body / validation_failed / meetings_not_configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "telegram_linked_to_other_account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/miniapp/meetings/{id}": {
      "patch": {
        "operationId": "miniappUpdateMeeting",
        "tags": [
          "miniapp"
        ],
        "summary": "Edit a single meeting (organizer/owner only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "this",
                "whole"
              ],
              "default": "this"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppMeetingUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated meeting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meeting"
                  ],
                  "properties": {
                    "meeting": {
                      "$ref": "#/components/schemas/MiniAppMeeting"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body / invalid time / meetings_recurring_unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "miniappDeleteMeeting",
        "tags": [
          "miniapp"
        ],
        "summary": "Cancel a single meeting (organizer/owner only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "this",
                "whole"
              ],
              "default": "this"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Meeting cancelled"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/miniapp/conflicts": {
      "post": {
        "operationId": "miniappConflicts",
        "tags": [
          "miniapp"
        ],
        "summary": "Cross-participant conflict warning",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppConflictsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conflicts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "occurrences"
                  ],
                  "properties": {
                    "occurrences": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MiniAppOccurrenceConflicts"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range / participants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "web_session"
      }
    },
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "MiniAppMeeting": {
        "type": "object",
        "required": [
          "id",
          "type",
          "dept",
          "host",
          "date",
          "start",
          "end",
          "rec",
          "organizer",
          "participants",
          "desc",
          "meet_link",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "dept": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "rec": {
            "type": "string"
          },
          "organizer": {
            "type": "string",
            "format": "email"
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            }
          },
          "desc": {
            "type": "string"
          },
          "meet_link": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "MiniAppMeetingsResponse": {
        "type": "object",
        "required": [
          "meetings"
        ],
        "properties": {
          "meetings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MiniAppMeeting"
            }
          }
        }
      },
...
      "MiniAppMeetingUpdateRequest": {
        "type": "object",
        "properties": {
          "dept": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "host": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "nullable": true,
            "description": "YYYY-MM-DD"
          },
          "start": {
            "type": "string",
            "nullable": true,
            "description": "HH:MM"
          },
          "end": {
            "type": "string",
            "nullable": true,
            "description": "HH:MM"
          },
          "desc": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "MiniAppConflictsRequest": {
        "type": "object",
        "required": [
          "participants",
          "date",
          "start",
          "end"
        ],
        "properties": {
          "participants": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            }
          },
          "date": {
            "type": "string",
            "description": "YYYY-MM-DD"
          },
          "start": {
            "type": "string",
            "description": "HH:MM Almaty"
          },
          "end": {
            "type": "string",
            "description": "HH:MM Almaty"
          },
          "exclude_id": {
            "type": "string",
            "format": "uuid"
          },
          "recurrence": {
            "type": "string",
            "enum": [
              "once",
              "daily",
              "weekly",
              "custom",
              "monthly"
            ]
          },
          "recurrence_until": {
            "type": "string",
            "format": "date"
          },
          "recurrence_days": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 7
            }
          }
        }
      },
      "MiniAppConflict": {
        "type": "object",
        "required": [
          "email",
          "name",
          "title",
          "start",
          "end"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "description": "HH:MM Almaty"
          },
          "end": {
            "type": "string",
            "description": "HH:MM Almaty"
          }
        }
      },
      "MiniAppOccurrenceConflicts": {
        "type": "object",
        "required": [
          "date",
          "start",
          "end",
          "conflicts"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "start": {
            "type": "string",
            "example": "10:00"
          },
          "end": {
            "type": "string",
            "example": "11:00"
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MiniAppConflict"
            }
          }
        }
      },
...
      "MiniAppMeetingCreateRequest": {
        "type": "object",
        "required": [
          "dept",
          "type",
          "date",
          "start",
          "end",
          "recurrence",
          "participants"
        ],
        "properties": {
          "dept": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "description": "YYYY-MM-DD"
          },
          "start": {
            "type": "string",
            "description": "HH:MM"
          },
          "end": {
            "type": "string",
            "description": "HH:MM"
          },
          "recurrence": {
            "type": "string",
            "enum": [
              "once",
              "daily",
              "weekly",
              "custom",
              "monthly"
            ]
          },
          "desc": {
            "type": "string"
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            }
          },
          "recurrence_until": {
            "type": "string",
            "format": "date"
          },
          "recurrence_days": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 7
            }
          }
        }
      },
...
    }
  }
}